making the transaction works!
Figured out the problem – I was passing in the buyerId instead of the buyer itself :(

But I fixed that. And then had to modify block.js because it would update the user's wealth for a second time (which is unnecessary). Fixed that, so now the make transaction form/function on the interface works well.

And now the getBlockchain part is also on the website, although it doesn't particularly work well.