updating the blockchain thing
before, when adding a new block, it wouldn't add a new block to the blockchain, so all of the transactions would be lost after the first block. So I fixed that. I also changed the number of transactions in each block to 5 (from 20) so that we wouldn't have to do as many transactions to show the extent of the simulation